refactor out new component
